Добавлено построение графика веса из данных базы данных

+++ app/src/main/java/com/example/dbvi/DBHelper.java
@@ -96,7 +96,7 @@ public class DBHelper extends SQLiteOpenHelper {
                         cursor.getInt(id_noSugar),
                         cursor.getInt(id_noSmoking),
                         cursor.getInt(id_eveningTeeth),
-                        cursor.getDouble(id_weight));
+                        cursor.getFloat(id_weight));
                 list.add(data);
             } while (cursor.moveToNext());
         db.close();
